Renaming updaters to capitalize and add EthFlow prefix
